% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{ell_basis}
\alias{ell_basis}
\title{The elliptic moment basis c_p, s_p}
\usage{
ell_basis(t, P, k)
}
\arguments{
\item{t}{angles in [0, 2*pi)}

\item{P}{highest order}

\item{k}{modulus}
}
\value{
list of two P-by-length(t) matrices
}
\description{
The elliptic moment basis c_p, s_p
}
\examples{
b <- ell_basis(seq(0, 2 * pi, length.out = 9), P = 2L, k = 0.9)
# at modulus zero the basis is the trigonometric one
max(abs(ell_basis(seq(0, 2 * pi, length.out = 9), 1L, 0)$C[1, ] -
        cos(seq(0, 2 * pi, length.out = 9))))
}
