% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{ell_KE}
\alias{ell_KE}
\title{Complete elliptic integrals of the first and second kind}
\usage{
ell_KE(k)
}
\arguments{
\item{k}{modulus, in [0, 1)}
}
\value{
list with components \code{K} and \code{E}
}
\description{
Complete elliptic integrals of the first and second kind
}
\examples{
ell_KE(c(0, 0.5, 0.9))
}
