% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{ell_moments}
\alias{ell_moments}
\title{Sample elliptic moments}
\usage{
ell_moments(t, P = 2L, k = 0.9)
}
\arguments{
\item{t}{angles}

\item{P}{highest order}

\item{k}{modulus}
}
\value{
P-by-2 matrix of (c, s) sample moments
}
\description{
Sample elliptic moments
}
\examples{
theta <- rellvm(200, mu = 1, kappa = 2, k = 0.9, seed = 1)
ell_moments(theta, P = 2L, k = 0.9)
}
