% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{ell_nome}
\alias{ell_nome}
\title{Nome q = exp(-pi K'/K)}
\usage{
ell_nome(k)
}
\arguments{
\item{k}{modulus}
}
\value{
A numeric vector of the same length as \code{k}, the nome for each modulus supplied.
}
\description{
Nome q = exp(-pi K'/K)
}
\examples{
ell_nome(c(0.5, 0.9))
}
