% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{fit_ellvm}
\alias{fit_ellvm}
\title{Minimum-distance fit of the elliptic von Mises}
\usage{
fit_ellvm(t, k = 0.9, P = 2L, ng = 512L)
}
\arguments{
\item{t}{angles}

\item{k}{modulus}

\item{P}{number of elliptic moments matched}

\item{ng}{quadrature points}
}
\value{
named vector of mu and kappa
}
\description{
Minimum-distance fit of the elliptic von Mises
}
\examples{
theta <- rellvm(500, mu = 1, kappa = 3, k = 0.9, seed = 7)
fit_ellvm(theta, k = 0.9, P = 2L)
}
