% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{dellvm}
\alias{dellvm}
\title{Elliptic von Mises density}
\usage{
dellvm(t, mu = 0, kappa = 1, k = 0.9, ng = 1024L)
}
\arguments{
\item{t}{angles}

\item{mu}{mean direction}

\item{kappa}{concentration}

\item{k}{modulus}

\item{ng}{quadrature points for the normalising constant}
}
\description{
Elliptic von Mises density
}
\examples{
theta <- seq(0, 2 * pi, length.out = 9)
dellvm(theta, mu = 1, kappa = 2, k = 0.9)
}
