% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{ell_dn2_moments}
\alias{ell_dn2_moments}
\title{Fourier coefficients of the arc-length family on an ellipse}
\usage{
ell_dn2_moments(k, P = 4L)
}
\arguments{
\item{k}{modulus}

\item{P}{number of harmonics}
}
\description{
On its own clock the family is dn^2/(4E); element one is E/K and element
element \code{n + 1} is \eqn{(2\pi^2/K^2) n q^n/(1-q^{2n})}.
}
\examples{
ell_dn2_moments(k = 0.8, P = 4L)
}
