% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellstat.R
\name{ell_jacobi}
\alias{ell_jacobi}
\title{Jacobi elliptic functions}
\usage{
ell_jacobi(u, k)
}
\arguments{
\item{u}{argument}

\item{k}{modulus}
}
\description{
Jacobi elliptic functions
}
\examples{
ell_jacobi(seq(0, 4, length.out = 5), k = 0.7)
}
